Logs why the import fails
"Installed" multiarray.so comes from the installer present on
osx105vmd locally.  "Built" multiarray.so comes from the build
directory which was *most likely* used to build the dmg.

Note that the dmg used for installation turned out to be smaller
than that which made it to sf.

The past of this py2.7 installers is rather unclear atm.